+++ tests/ui/specs/WizardPanels/WizardPanels-specs.js
@@ -98,6 +98,23 @@ describe('WizardPanels', function () {
 	});
 
 	describe('Focus Behavior', function () {
+		it('should focus a component within the body on mount', function () {
+			wizardPanels.waitForEnter(1);
+
+			const expected = 'OK';
+			const actual = browser.execute(getFocusedTextContent);
+
+			expect(actual).to.be.equal(expected);
+		});
+
+		it('should focus a navigation button on mount if no focusable elements exist in the body', function () {
+			Page.open('?index=3');
+
+			wizardPanels.waitForEnter(4);
+
+			expect(wizardPanels.prevButton.isFocused()).to.be.true();
+		});
+
 		it('should select contents over buttons - [GT-29594]', function () {
 			wizardPanels.focusNextButton();
 			Page.spotlightSelect();